Refactoring ConvolutionParameters and added documentation
